+++ library/src/schemas/nullish/nullishAsync.ts
@@ -21,7 +21,7 @@ export type NullishSchemaAsync<
 > = BaseSchemaAsync<Input<TWrapped> | null | undefined, TOutput> & {
   schema: 'nullish';
   wrapped: TWrapped;
-  get default(): TDefault;
+  getDefault(): TDefault;
 };
 
 /**
@@ -54,9 +54,9 @@ export function nullishAsync<
     wrapped,
 
     /**
-     * The default value.
+     * Retutns the default value.
      */
-    get default() {
+    getDefault() {
       return typeof default_ === 'function'
         ? (default_ as () => TDefault)()
         : (default_ as TDefault);
@@ -80,7 +80,7 @@ export function nullishAsync<
       let default_: Awaited<TDefault>;
       const value =
         (input === null || input === undefined) &&
-        (default_ = await this.default) &&
+        (default_ = await this.getDefault()) &&
         default_ !== undefined
           ? default_
           : input;