+++ library/src/schemas/optional/optional.ts
@@ -13,7 +13,7 @@ export type OptionalSchema<
 > = BaseSchema<Input<TWrapped> | undefined, TOutput> & {
   schema: 'optional';
   wrapped: TWrapped;
-  get default(): TDefault;
+  getDefault(): TDefault;
 };
 
 /**
@@ -43,9 +43,9 @@ export function optional<
     wrapped,
 
     /**
-     * The default value.
+     * Returns the default value.
      */
-    get default() {
+    getDefault() {
       return typeof default_ === 'function'
         ? (default_ as () => TDefault)()
         : (default_ as TDefault);
@@ -66,7 +66,7 @@ export function optional<
      */
     _parse(input, info) {
       // Get default or input value
-      const value = input === undefined ? this.default : input;
+      const value = input === undefined ? this.getDefault() : input;
 
       // Allow `undefined` value to pass
       if (value === undefined) {